utils/requirements.txt
gitdb==4.0.12
GitPython==3.1.46
smmap==5.0.2